Consolidate the 32-bit and 64-bit config files into one
There was only one small difference between the two and now that we are
tracking 32-bit vs 64-bit in our post scripts it is easy enough to make
the mod in scripting.

This also makes some small config changes:
* Increase the gpu_mem for our 4GB mem system.
* Remove the unnecessary dtc overlay for the old USB controller since
otg_mode is all that we need for USB to work.