Update dev certificates; **need to flash the update**
This changes the dev certificates to give them longer length.  The set
we had expired in about 10 years.  The new ones expire in 50 years.

Since the certs changed you need to flash this change to the device
instead of using an OTA update.

Also I changed the update script to just show the entire status blob on
an update failure so the user can see the RAUC error better.